data_pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alpaca-py>=0.42.2",
  "httpx>=0.28.1",
  "newsapi-python>=0.2.7",
  "praw>=7.8.1",
  "prefect>=3.4.23",
  "pydantic>=2.12.2",
  "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
  "ruff>=0.14.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
